move word complexity assessment
