+++ packages/yoastseo/src/scoring/assessments/premiumAssessments/readability/WordComplexityAssessment.js
@@ -1,11 +1,11 @@
 import { __, sprintf } from "@wordpress/i18n";
 import { merge } from "lodash-es";
 
-import { collectMarkingsInSentence } from "../../../languageProcessing/helpers/word/markWordsInSentences";
-import { createAnchorOpeningTag } from "../../../helpers/shortlinker";
-import AssessmentResult from "../../../values/AssessmentResult";
-import Mark from "../../../values/Mark";
-import Assessment from "../assessment";
+import { collectMarkingsInSentence } from "../../../../languageProcessing/helpers/word/markWordsInSentences";
+import { createAnchorOpeningTag } from "../../../../helpers/shortlinker";
+import AssessmentResult from "../../../../values/AssessmentResult";
+import Mark from "../../../../values/Mark";
+import Assessment from "../../assessment";
 
 /**
  * Represents the assessment that checks whether there are too many complex words in the text.
