+++ packages/yoastseo/src/scoring/storePostsAndPages/cornerstone/contentAssessor.js
@@ -8,7 +8,7 @@ import TransitionWords from "../../assessments/readability/TransitionWordsAssess
 import PassiveVoice from "../../assessments/readability/PassiveVoiceAssessment.js";
 import SentenceBeginnings from "../../assessments/readability/SentenceBeginningsAssessment.js";
 import TextPresence from "../../assessments/readability/TextPresenceAssessment.js";
-import WordComplexityAssessment from "../../assessments/readability/WordComplexityAssessment";
+import WordComplexityAssessment from "../../assessments/premiumAssessments/readability/WordComplexityAssessment";
 
 /*
  Temporarily disabled:
